mobile friendly home page
Began mobile work for home page. Unfinished, currently an extra class
with extra css. Best practice would be to find a way to turn off the
carousel, but for time constraints I went the other direction for now to
get it working and then turn off the classes in the future.